cefit_manuales/requirements.txt
mysql-connector-python
bcrypt
python-dotenv
Pillow
PyMuPDF
PyPDF2
Pyinstaller
cx_Freeze
